+++ Acadar/app/_layout.js
@@ -1,5 +1,6 @@
 import { Slot } from 'expo-router';
-import {useState, createContext } from 'react';
+import {useState, createContext} from 'react';
+import * as SecureStore from 'expo-secure-store';
 const UserContext = createContext();
 
 export default function Root(){
@@ -11,10 +12,11 @@ export default function Root(){
         setField(category);
     }
     async function add(userField, value){
-        var body1 = await fetch(`http://localhost:5000/functions/add/${userField}`, {
+        const token = await SecureStore.getItemAsync("authToken");
+        if(!token) return;
+        var body1 = await fetch(`http://172.233.131.223:5000/functions/add/${userField}`, {
             method: 'POST',
             headers: { "Content-Type": "application/json" },
-            credentials: "include",
             body: JSON.stringify(value)
         });
         var body1_response = await body1.json();
@@ -23,8 +25,10 @@ export default function Root(){
         }
     };
     async function remove(userField, name){
+        const token = await SecureStore.getItemAsync("authToken");
+        if(!token) return;
         if(name.includes(" ")) name = name.split(" ").join("%20")
-        var body1 = await fetch(`http://localhost:5000/functions/remove/${userField}/${name}`, {
+        var body1 = await fetch(`http://172.233.131.223:5000/functions/remove/${userField}/${name}`, {
             method: 'DELETE',
             headers: { "Content-Type": "application/json" },
             credentials: "include"
@@ -35,9 +39,11 @@ export default function Root(){
         }
     };
     async function update(userField, name){
+        const token = await SecureStore.getItemAsync("authToken");
+        if(!token) return;
         if(value.trim() === "") return;
         if(name.includes(" ")) name = name.split(" ").join("%20")
-        var body1 = await fetch(`http://localhost:5000/functions/update/${userField}/${name}/${field}`, {
+        var body1 = await fetch(`http://172.233.131.223:5000/functions/update/${userField}/${name}/${field}`, {
             method: 'PUT',
             headers: { "Content-Type": "application/json" },
             credentials: "include",
